✨ feat: deleteItem 예외 처리 작성

+++ src/main/java/com/project/vinpong/service/ItemService/ItemCommandServiceImpl.java
@@ -108,9 +108,11 @@ public class ItemCommandServiceImpl implements ItemCommandService{
     }
 
     @Override
-    public void deleteById(Long itemId) {
+    @Transactional
+    public void deleteById(Long itemId, String username) {
         Optional<Item> item = itemRepository.findById(itemId);
         if (item.isEmpty()) throw new ItemHandler(ErrorStatus.ITEM_NOT_FOUND);
+        if (!item.get().getSeller().getUsernamae().equals(username)) throw new ItemHandler(ErrorStatus.SELLER_NOT_MATCHED);
         itemRepository.deleteById(itemId);
     }
 }
